requirements.txt
ahrs==0.3.1
# azure_storage==0.37.0
gps==3.19
Mock==4.0.3
numpy==1.19.5
paho_mqtt==1.6.1
pyserial==3.5
pytz==2021.3
PyYAML==6.0
pyzmq==23.1.0
quaternion==3.5.2.post4
scipy==1.8.1
smbus==1.1.post2
uptime==3.0.1
